requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio @ file:///home/compiling-ganesh/24m0797/wheels/python310/anyio-4.12.0-py3-none-any.whl
apache-tvm-ffi==0.1.8.post2
astor==0.8.1
async-timeout==5.0.1
attrs==25.4.0
blake3==1.0.8
cachetools==6.2.4
cbor2==5.8.0
certifi @ file:///home/compiling-ganesh/24m0797/wheels/python310/certifi-2025.11.12-py3-none-any.whl
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.12.2
cryptography==46.0.3
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup @ file:///home/compiling-ganesh/24m0797/wheels/python310/exceptiongroup-1.3.1-py3-none-any.whl
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.20.3
flashinfer-python==0.5.3
frozenlist==1.8.0
fsspec==2026.1.0
gguf==0.17.1
h11 @ file:///home/compiling-ganesh/24m0797/wheels/python310/h11-0.16.0-py3-none-any.whl
hf-xet==1.2.0
httpcore @ file:///home/compiling-ganesh/24m0797/wheels/python310/httpcore-1.0.9-py3-none-any.whl
httptools==0.7.1
httpx @ file:///home/compiling-ganesh/24m0797/wheels/python310/httpx-0.28.1-py3-none-any.whl
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna @ file:///home/compiling-ganesh/24m0797/wheels/python310/idna-3.11-py3-none-any.whl
ijson==3.4.0.post0
interegular==0.3.3
jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.25.0
mdurl==0.1.2
mistral-common==1.8.8
model-hosting-container-standards==0.1.13
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.17.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.3.5
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.15.0
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
outlines-core==0.2.11
packaging==25.0
partial-json-parser==0.2.1.1.post7
pillow==12.1.0
pip @ file:///home/conda/feedstock_root/build_artifacts/pip_1762776163052/work
prometheus-client==0.24.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
py-cpuinfo==9.0.0
pybase64==1.4.3
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pysocks @ file:///home/compiling-ganesh/24m0797/wheels/python310/PySocks-1.7.1-py3-none-any.whl
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pyyaml==6.0.3
pyzmq==27.1.0
ray==2.53.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.49.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
socksio @ file:///home/compiling-ganesh/24m0797/wheels/python310/socksio-1.0.0-py3-none-any.whl
sse-starlette==3.1.2
starlette==0.50.0
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.6
triton==3.5.0
typer==0.21.1
typing-extensions @ file:///home/compiling-ganesh/24m0797/wheels/python310/typing_extensions-4.15.0-py3-none-any.whl
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vllm==0.13.0
watchfiles==1.1.1
websockets==16.0
wheel==0.45.1
xgrammar==0.1.27
yarl==1.22.0

# --- vcoder training dependencies ---
trl>=0.28.0
peft>=0.18.1
accelerate
datasets
tensorboard
qwen-vl-utils
playwright
scikit-image
beautifulsoup4
html5lib
lxml
